we all like correctness + faster + laziness, so we got it all!
correctness: do not do the join of all audio tables, so we still get
those music without albums, artist or genres.

faster: this is also faster, as there is less data comming in, and
during db_song_copy() we do not have to handle album, artist and genre
strings.

laziness: do just when item requests such information. So if you open
the list but never scrolls, just a few will be fetched. And the
information is saved for future reuse.




SVN revision: 52664