improve enjoy theme.
bit more scale and finger_size friendly, but there are other things to do;

   - add missing elm.swallow.pad to all genlist items (test with finger size)
   - ensure scale works properly
   - check/fix now playing screen (not even looked yet)

I have no idea why, but genlists are not updating dynamically. It does
work with elementary_test, but not within enjoy, so probably some
widget is broken wrt propagating theme changes to children. To test
you need to exit the list and enter it again.




SVN revision: 53827